Fix `{foreach}` loops when the to-be-iterated value is empty
The restoring of the `item` value failed, because the necessary data in
`foreachVars` was only being filled when the `foreach()` loop was actually
entered. Move this saving of the old value up to ensure it always happens.

see 75ce18bc18904d1215c7d021ac0ac18c0a7a5d42
see #4425