Take the array key into account when checking whether a FOREIGN KEY is up to date in DatabaseTableChangeProcessor
Previously updating a FOREIGN KEY that looks like this:

    […] FOREIGN KEY (fooID) REFERENCES wcf1_foo (fooID) ON DELETE SET NULL ON UPDATE CASCADE

to:

    […] FOREIGN KEY (fooID) REFERENCES wcf1_foo (fooID) ON DELETE CASCADE ON UPDATE SET NULL

would not do anything.

Converted into the `getData()` representation of the PHP DDL API these would
look identical when looking at the values only. Both span the same column and
reference the same column on the same table. The `ON …` actions are one
`CASCADE` and one `SET NULL` for both.

For this reason the diff is empty, believing that the FOREIGN KEY already
matches the expected configuration.

Fix this issue by using `array_diff_assoc` which will also take the key into
account.