Increase defaultvalue of user_cleanup_notification_lifetime to 31 days
A notification lifetime of just 2 weeks is insufficient, because it might
easily result in an user losing important notifications over their summer
vacation.

As an example: If a user checks out on a Friday afternoon and checks in on a
Monday morning 2 weeks later, they'll have lost any notifications arriving on
the first weekend weekend. If their vacation is even longer, e.g. due to a
public holiday, they'll also lose non-weekend notifications.

Increase the lifetime to 31 days. This spans a range from a Friday afternoon
until the Monday afternoon 4 weeks later. This should be sufficiently long for
pretty much all types of vacation.

The old default of 14 days dates back to December 2014 in commit
5cdf8c0338381d1c880bd07d46bb7fcbbde09b61.