Remove return value for AbstractSearchIndexManager::createSearchIndex()
Returning this boolean value does not appear to be useful at all, as there is
no reason why the state after this method finishes should be that the INDEX
does not actually exist (except in case of an Exception). Whether or not it
previously existed is irrelevant.

In fact this method is `protected` and the return value is not used at all,
thus it is safe to remove this requirement.