Remove link from requestURI in acpSessionLog
The link has been non-functional since ages, because `hasProtectedURI()`
effectively always returned `true`, as the `page` and `form` query parameters
are no longer in use.

With WoltLab Suite 5.4 the `requestURI` might also contain a request ID and the
link length is also limited to 255 characters.

All in all it appears that supporting a link within the log is non-trivial,
with the benefit being rather small.

Resolves #4524