Add UploadHandler session cleanup
Since the sessions stick way longer than with the introduction of the system, it is necessary, to clean up old file uploads in the session storage.

With this patch, a field in the session saves the creation time of the StorageField in the session and delete the field after 24 hours automatically.