implement optional forcing decimals in `StringUtil.formatNumeric()`
With this change the current behavior of `StringUtil.formatNumeric()` is kept als default behavior, but it is possible to force the given number of decimals instead of treating the parameter as `$maxDecimals`.
This change makes it possible to get the same result as using `CurrencyModifierTemplatePlugin` in order to display things in a more consistent way.

see https://www.woltlab.com/community/thread/294033/