Make CacheHandler::getCacheName() less expensive
The old implementation generated cache filenames that are a little cleaner to
the human eye, but the generation was needlessly expensive, even calling the
UTF-8-safe StringUtil::firstCharToUpperCase() on an ASCII value.

It will need to be checked if the changing cache name will cause issues during
the upgrade to 5.5, as it will implicitly result in a cache clear.