Add `IUniqueNameXMLPackageInstallationPlugin` to enforce unique names
Debugging is very tedious when elements with non-unique names are supplied in a plugin, because the first element is always overwritten with the second element. This can never be intentional and is always a fault of the developer, which is difficult to find. This interface ensures that all identifiers supplied for a PIP are unique.