Embed the assets as data URIs in rescueMode.tpl
The same reasoning as for the change to WCFSetup in
600940bc4fe0cbc219ba6fd9a89444aefa07dda5 applies:

With this change the rescue mode is fully self-contained and does not rely on
external requests that need to be served using PHP, making it more resilient.

It also avoids the expensive rebuild of the compiled ACP CSS, because it now
uses the prebuilt WCFSetup.css which can be used as-is for this purpose.