Disable hiding of the user menu on Redactor focus
The behavior was intended to increase the available screen space when the editor is active. However, the height of smartphones has constantly increased, reducing the impact of a couple extra pixels.

The major downside was the focus behavior which is somewhat unintuitive on mobile and has repeatedly caused confusion among users, because the menu was seemingly “lost”.