Remove unneeded usage of `mb_strpos`
If the return value of `mb_strpos` is only compared to `false` then there is no
need to use the multibyte engine, instead `str_contains()` does example the
same and is clearer.

The same applies if the return value is compared to `0`, in this case the size
of multibyte characters cannot have affected the offset. `str_starts_with()`
can be used instead.