Fix incorrect use of `mb_strpos` in MysqlSearchEngine
The return value of `mb_strpos` needs to be checked type-safely, as both `0`
and `false` are falsy. In this case this likely was safe, as the inner join may
not appear at the start of the query, it was a questionable nonetheless.

Fix this by using `str_contains()` which makes the intent even clearer.