Always use install.php's __DIR__ as the temporary directory
The previous implementation of `BasicFileUtil::getTempFolder()` was unstable,
as it automatically detected `tmp/` directories in the document root. If
WoltLab Suite Core is installed into the document root, then the extraction of
the files will implicitly create a `tmp/` directory and thus change the temp
folder for the remaining installation.