Fix code style in WCFACP
Apparently PSR-12 allows either brace placement, with PHP-CS-Fixer preferring
the brace on the same line and Intelephense preferring it on the next line. If
a multi-line interface definition would be used here, then the brace must be on
a dedicated line, making Intelephense slightly more right on this one.

However as WCFACP is expected to be touched seldomly, thus seldomly running the
Intelephense formatter, the pragmatic solution is to satisfy PHP-CS-Fixer.

see FriendsOfPHP/PHP-CS-Fixer#6396 for a related issue