Remove Strato workaround from WCFSetup
This workaround exists since the very first commit and it's not clear why that
was needed back then. It likely was related to the temporary directory not
being writable or files not being deletable in some cases, but with the change
of WCFSetup's temporary directory anything less than full writing permissions
to the directory mean that WoltLab Suite won't actually run after installation
anyway, making this workaround obsolete.

Also the fact that decompression of the archive does not happen in all cases
implies that the decompression is not actually necessary at this stage, making
that the next candidate for removal.