Moved the notification list into the global space
The notification list was placed in the user menu a long time ago, but never really fitted in there. The link was already (and continues to be) reachable through the drop-down menu, removing the need of an additional location.

Moving the notification list into the global page space makes more sense and aligns with similar lists such as the conversations.