Remove enforced redirect of `login` and `index` controller in ACP to the `wcf` application
It is not clear why this was added in d49006fac289699c925a6d5644f102b7ebfc972c,
but it does not appear to still be required. The login check in WCFACP
correctly redirects to the `wcf` app even if an application controller is
requested. The index controller redirect was cleaned up in the previous commit.