Simplify SetupLanguage::loadLanguage()
The caching of the parsed and compiled phrases adds quite a bit of complexity
for little gain.

After this change the `loadLanguage()` method takes ~10ms on my computer which
is acceptable for WCFSetup which is not performance critical and where this
time is absolutely dwarfed by the other installation processes, such as files
extraction.

This also uncovered a pretty hard to debug issue: LanguageFactory must not be
initialized early in WCFSetup, because the cache data will simply be empty (due
to the lack of database access). This was previously hidden by
SetupLanguage::loadLanguage() being a noop if the compiled data existed, thus
masking the problem for subsequent requests.