Match foreign keys independent of referenced column casing in PHP DDL
Early versions of MySQL 8 contained a bug that would return the column names of
the referenced table in lowercase whenever querying the FOREIGN KEY or when
showing the table's structure:

https://bugs.mysql.com/bug.php?id=88718

This issue also affects the any created database dumps and is not corrected
when importing the dump into a fixed version of MySQL:

https://bugs.mysql.com/bug.php?id=98976

Thus any instances that were ever dumped with an affected version of MySQL 8
will have foreign keys with a mismatching column case.

The FOREIGN KEY itself will be completely functional, MySQL does not care about
the casing of column names.

However the PHP DDL considers those FOREIGN KEYs to be different when
attempting to find an existing FOREIGN KEY to update, thus believing that the
FOREIGN KEY does not yet exist. As a result it attemtps to create it,
generating a name that is identical to the existing FOREIGN KEY's, leading to
conflicts when attempting to log the change into package_installation_sql_log.

Fix this issue by lowercasing the referenced column names when grabbing the
`->getDiffData()`. The PHP DDL will then find the existing FOREIGN KEY, but it
will realize that it is different when diffing the `->getData()`.

This will lead to the "broken" FOREIGN KEY being dropped and it being recreated
correctly, both avoiding the duplicate key error and also actually correcting
the column name.