Support returning responses in RequestHandler::buildRequest()
The return type of `->buildRequest()` is a little ugly now, but this will be
better than a hard call to `exit;` whenever a redirect is desired.

This whole routing and request building logic would ideally move into a
Middleware, this can be considered a first step to clean that up.